fashionbackend/fashionbackendapp/requirements.txt
altgraph==0.17.3
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
blessed==1.20.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
dj-database-url==2.3.0
Django==4.2.19
django-cors-headers==4.7.0
django-environ==0.12.0
django-picklefield==3.3
django-q==1.3.9
django-restframework==0.0.1
django-shortcuts==1.6
djangorestframework==3.15.2
firebase-admin==6.7.0
future==0.18.3
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
joblib==1.4.2
kombu==5.4.2
macholib==1.16
msgpack==1.1.0
numpy==2.0.2
pillow==11.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==3.5.3
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
six==1.16.0
sqlparse==0.5.3
threadpoolctl==3.5.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
